requirements.txt
fastapi
uvicorn
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters
langchain-huggingface
sentence-transformers
transformers
faiss-cpu
gtts
openai-whisper
python-multipart
huggingface_hub
torch
python-jose
python-dotenv
